pyproject.toml
[project]
name = "robotframework-construct"
description = "Infrastructure to use construct Structs in robotframework"
authors = [
    {name = "Franz Haas", email = "franz.dominik.haas@gmail.com"},
]
dynamic = ["version"]

readme = "README.md"
requires-python = ">=3.10.0"
dependencies = [
    "robotframework>=7.1",
    "construct==2.10.69",
]
license = "Apache-2.0"
keywords = ["robotframework"]

[project.urls]
Documentation = "https://marketsquare.github.io/robotframework-construct"
Repository = "https://github.com/MarketSquare/robotframework-construct"
Release_Notes = "https://marketsquare.github.io/robotframework-construct/release_notes.md"

[project.optional-dependencies]
test = [
    "cosmic-ray>=8.3.15",
    "pymongo>=4.9.1",
    "ruff>=0.7.2",
    "pytest>=8.3.3",
]
nci = ["pyserial"]
breakout = [
    "ruamel_yaml>=0.18.6",
    "xonsh>=0.18.4",
]

[build-system]
requires = [
    "hatchling",
    "versioningit",
]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "versioningit"

[tool.uv]
dev-dependencies = [
    "ipython>=8.29.0",
    "radon>=6.0.1",
    "ruff>=0.7.2",
    "ruamel_yaml>=0.18.6",
    "mypy>=1.13.0",
    "xonsh>=0.18.4",
]

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "node-and-timestamp"
